Visualise and inspect neural net
Adds the ability to dump an activated neural net to json, load it again
from json and graph it with graphviz.

Generating a graphviz image is slow, these images are large clocking in
at ~200mb. You may not want to do this during a trainning run so we
support storing the data required for generating an image to use at a
later date.

Usage:
```rb
  # make a prediction to set the activations
  neural_net.predict(example.image_bytes)
  # Store the activated neural net and weights to disk
  Graphs::Graph.new(neural_net).save("tmp/some-activation.json")
  # Load the activations and weights
  graph = Graphs::Graph.load("tmp/some-activation.json")
  # Genereate a (large) png of the network
  Graphs::Graphviz.new(graph).to_png("tmp/activations.png")
```

Might be useful for visualising how the weights change as the network
learns or seeing what the average activations are for a given digit.